changed int declarations to int8_t/int16_t/int32_t/uint8_t/uint16_t/uint32_t to meet MISRA standards

+++ inc/mode_default.h
@@ -8,7 +8,7 @@ enum mode_power_source {
 };
 
 namespace mode_default {
-    int run(double, double, bool);
+    uint8_t run(double, double, bool);
 }
 
 #endif
