+++ inc/modes.h
@@ -2,17 +2,17 @@
 #define MODES_H
 
 namespace device {
-    void set_mode(int);
-    int get_mode();
+    void    set_mode(uint8_t);
+    uint8_t get_mode();
     
-    void run(int);
+    void    run(uint8_t);
 }
 
 //test functions. Delete for release
 namespace device_tests {
-    void set_temperature_max(double);
-    void set_temperature_current(double);
-    void set_pv_available(bool);
+    void    set_temperature_max(double);
+    void    set_temperature_current(double);
+    void    set_pv_available(bool);
 }
 
 #endif
